+++ app/Icons/page.tsx
@@ -3,8 +3,8 @@ import { Container } from "react-bootstrap";
 import IconHeader from "../WebsiteHeaders/icons_header";
 import IconsDescription from "../WebsiteHeaders/icons_description";
 import IconsTutorial from "../WebsiteHeaders/icons_tutorial";
-import ScrollingIcons from "../WebsiteComponents/scrollingIcons";
-import IconPayment from "../WebsiteComponents/icons_payment";
+import ScrollingIcons from "../WebsiteComponents/IconsComponents/scrollingIcons";
+import IconPayment from "../WebsiteComponents/IconsComponents/icons_payment";
 import PaymentSuccess from "../WebsiteComponents/PaymentForm/PaymentSuccess";
 
 export default function Icons() {
