+++ app/WebsiteComponents/Education/education_component.tsx
